Add kilde and reorganize directories
