+++ src/components/forms/vedtak/Vedtak.tsx
@@ -5,24 +5,24 @@ import { useMutation, useQuery } from "@tanstack/react-query";
 import React from "react";
 import { useParams } from "react-router-dom";
 
-import { Grunnlag, ResultatPeriode, RolleDtoRolleType } from "../../api/BidragBehandlingApi";
-import { OpprettGrunnlagRequestDto } from "../../api/BidragVedtakApi";
-import { BEHANDLING_API, BIDRAG_VEDTAK_API } from "../../constants/api";
-import { useForskudd } from "../../context/ForskuddContext";
-import { Avslag } from "../../enum/Avslag";
-import environment from "../../environment";
-import { useGetBehandling, usePersonsQueries } from "../../hooks/useApiData";
+import { Grunnlag, ResultatPeriode, RolleDtoRolleType } from "../../../api/BidragBehandlingApi";
+import { OpprettGrunnlagRequestDto } from "../../../api/BidragVedtakApi";
+import { BEHANDLING_API, BIDRAG_VEDTAK_API } from "../../../constants/api";
+import { useForskudd } from "../../../context/ForskuddContext";
+import { Avslag } from "../../../enum/Avslag";
+import environment from "../../../environment";
+import { useGetBehandling, usePersonsQueries } from "../../../hooks/useApiData";
 import {
     mapBehandlingReferanseliste,
     mapGrunnlagPersonInfo,
     mapResultatKodeToDisplayValue,
-} from "../../mapper/VedtakBeregningkMapper";
-import { uniqueByKey } from "../../utils/array-utils";
-import { toISODateTimeString } from "../../utils/date-utils";
-import { FlexRow } from "../layout/grid/FlexRow";
-import { PersonNavn } from "../PersonNavn";
-import { QueryErrorWrapper } from "../query-error-boundary/QueryErrorWrapper";
-import { RolleTag } from "../RolleTag";
+} from "../../../mapper/VedtakBeregningkMapper";
+import { uniqueByKey } from "../../../utils/array-utils";
+import { toISODateTimeString } from "../../../utils/date-utils";
+import { FlexRow } from "../../layout/grid/FlexRow";
+import { PersonNavn } from "../../PersonNavn";
+import { QueryErrorWrapper } from "../../query-error-boundary/QueryErrorWrapper";
+import { RolleTag } from "../../RolleTag";
 
 function grunnlagTilOpprettGrunnlagRequestDto(grunnlag: Grunnlag): OpprettGrunnlagRequestDto {
     return {
