+++ src/enum/KildeTexts.ts
@@ -1,0 +1,4 @@
+export enum KildeTexts {
+    OFFENTLIG = "Offentlig",
+    MANUELT = "Manuelt",
+}
